chore: update code style and make util function
